% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{add.HiTMetadata}
\alias{add.HiTMetadata}
\title{Add metadata to an annotated Seurat object to then obtain a HiT object}
\usage{
add.HiTMetadata(
  object = NULL,
  scGate.models = NULL,
  add.signatures = NULL,
  layer2_annotation = "functional.cluster",
  ref.maps = NULL,
  layer2_levels = NULL
)
}
\arguments{
\item{object}{Annotated Seurat object}

\item{scGate.models}{names of the scGate models used for layer1 annotation. Usually this is a list, here names of the list should be provided.}

\item{add.signatures}{names of the signatures computed during HiTME annotation, provide names of the list of signatures.}

\item{layer2_annotation}{Metadata column with layer2 annotation data.}

\item{ref.maps}{Reference maps used for annotation on layer2 can be provided to extract their information.}

\item{layer2_levels}{If reference maps are not provided, a vector or list with the levels of the layer2 annotation can be provided. These must be named with corresponding layer1_link.}
}
\value{
Reference compatible with HiTME annotation, linking layer1 and layer2.
}
\description{
Add metadata in misc slot of a Seurat object previously annotated to run \link{get.HiTObject}.
This needed data is by default incorporated by \link{Run.HiTME}, however, in objects annotated otherwise or splitted or merged these slots are lost. This metadata is needed basically to be able to link layer1 and layer2.
}
