% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{annotate_cells}
\alias{annotate_cells}
\title{Annotate cell types in parallel}
\usage{
annotate_cells(dir, scGate.model, ref.maps, ncores = 1, progressbar = TRUE)
}
\arguments{
\item{dir}{The directory where your sample .rds files(seurat objects) are located}

\item{scGate.model}{The scGate model to use (use get_scGateDB() to get a list of available models)}

\item{ref.maps}{A named list of the ProjecTILs reference maps to use}

\item{ncores}{The number of cores to use}

\item{progressbar}{Whether to show a progressbar or not}
}
\value{
No return in R. The input files will be overwritten.
}
\description{
Annotate cell types in parallel
}
\examples{
library(ProjecTILs)

path_data <- file.path("~", "Dropbox", "CSI", "Standardized_SingleCell_Datasets", "ZhangY_2022_34653365", "output", "test")

# Define scGate model
scGate_models_DB <- scGate::get_scGateDB(branch = "master", verbose = T, force_update = TRUE)
models.TME <- scGate_models_DB$human$TME_HiRes

# Load ProjecTILs reference maps
path_ref <- "~/Dropbox/CSI/reference_atlases"
ref.maps <- list(CD8 = load.reference.map(file.path(path_ref, "CD8T_human_ref_v1.rds")),
                 CD4 = load.reference.map(file.path(path_ref, "CD4T_human_ref_v2.rds")),
                 DC = load.reference.map(file.path(path_ref, "DC_human_ref_v1.rds")),
                 MoMac = load.reference.map(file.path(path_ref, "MoMac_human_v1.rds")))
annotate_cells(dir = path_data,
               scGate.model = models.TME,
               ref.maps = ref.maps,
               ncores = 6)
}
