% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calc_CTcomp}
\alias{calc_CTcomp}
\title{Calculate cell type composition}
\usage{
calc_CTcomp(
  object,
  sample.col = NULL,
  annot.cols = "scGate_multi",
  min.cells = 10,
  useNA = FALSE,
  rename.Multi.to.NA = FALSE
)
}
\arguments{
\item{object}{A seurat object}

\item{sample.col}{The Seurat object metadata column containing sample names}

\item{annot.cols}{The Seurat object metadata column(s) containing celltype annotations (provide as character vector, containing the metadata column name(s))}

\item{min.cells}{Set a minimum threshold for number of cells to calculate relative abundance (e.g. less than 10 cells -> no relative abundnace will be calculated)}

\item{useNA}{Whether to include not annotated cells or not (labelled as "NA" in the annot.cols). Can be defined separately for each annot.cols (provide single boolean or vector of booleans)}

\item{rename.Multi.to.NA}{Whether to rename cells labelled as "Multi" by scGate to "NA"}
}
\value{
Cell type compositions as a list of data.frames containing cell counts, relative abundance (freq) and clr-transformed freq (freq_clr), respectively.
}
\description{
Calculate cell type composition
}
\examples{
devtools::install_github('satijalab/seurat-data')
                         library(SeuratData)
                         options(timeout = max(300, getOption("timeout")))
                         InstallData("panc8")
                         data("panc8")
                         panc8 = UpdateSeuratObject(object = panc8)
# Calculate overall composition
celltype.compositions.overall <- calc_CTcomp(object = panc8, annot.cols = "celltype")

# Calculate sample-wise composition
celltype.compositions.sample_wise <- calc_CTcomp(object = panc8, annot.cols = "celltype", sample.col = "orig.ident")

# Plot overall composition
par(mar = c(8.1, 4.1, 4.1, 2.1))
barplot(unlist(celltype.compositions.overall[["celltype"]][["freq"]]),
        ylab = "Relative abundance (\%)",
        las = 2)
# Plot sample-wise composition
df <- celltype.compositions.sample_wise[["celltype"]][["freq"]]
dflong <- reshape2::melt(t(df))
library(ggplot2)
ggplot(dflong, aes(x=Var1, y=value, color=Var1)) +
  geom_boxplot() + geom_point() + theme_classic() +
  theme(axis.text.x = element_text(angle = 45, vjust = 1, hjust=1)) + labs(x = "", y = "Relative abundance (\%)") + NoLegend()
}
