% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{composition.barplot}
\alias{composition.barplot}
\title{Render bar plots summarizing celltype proportions and distribution in samples and groups}
\usage{
composition.barplot(
  hit.object = NULL,
  sample.col = NULL,
  layer = "layer1",
  return.plot.to.var = FALSE,
  facet.by = NULL
)
}
\arguments{
\item{hit.object}{A Hit class object (typically after applying merge.HiTObjects onto a list of HiTObjects)}

\item{layer}{Default "layer1" if you have one cell type annotation layer in your hit.object. Alternatively "layer2" etc. if you have multiple layers of annotation depths.}

\item{return.plot.to.var}{Optionally, you can save the ggplots to a variable if you would like to further modify and adapt the plots on your own.}

\item{facet.by}{This allows you to pass a metadata column name present in your hit.object$metadata to show your samples in facets with ggplot facet_grid, for example by "condition".}
}
\value{
Plotting function to show the cell type composition from HiTME object across different samples.
}
\description{
Render bar plots summarizing celltype proportions and distribution in samples and groups
}
