% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{composition.boxplot}
\alias{composition.boxplot}
\title{Render box plots summarizing celltype proportions and distribution in samples and groups}
\usage{
composition.boxplot(
  hit.object = NULL,
  plot.var = "clr",
  layer = "layer1",
  return.plot.to.var = FALSE,
  group.by = NULL,
  facet.by = NULL,
  pval.method = "wilcox.test",
  p.adjust.method = "BH",
  palette = "lancet",
  legend.position = "right"
)
}
\arguments{
\item{hit.object}{A Hit class object (typically after applying merge.HiTObjects onto a list of HiTObjects)}

\item{plot.var}{Column in the hit.object$composition: either "freq" for cell type relative abundance in percent or "clr" (for centered log-ratio transformed). Default: "clr" as it is better suited for statistical analysis and is better able to also show low abundant cell types.}

\item{layer}{Default "layer1" if you have one cell type annotation layer in your hit.object. Alternatively "layer2" etc. if you have multiple layers of annotation depths.}

\item{return.plot.to.var}{Optionally, you can save the ggplots to a variable if you would like to further modify and adapt the plots on your own.}

\item{group.by}{This allows you to pass a metadata column name present in your hit.object$metadata to show your samples in groups, for example by "condition".}

\item{facet.by}{This allows you to pass a metadata column name present in your hit.object$metadata to show your samples in facets with ggplot facet_grid, for example by "condition".}

\item{pval.method}{Specify method how to calculate the p-value. Wilcoxon test is recommended as compositional data might not fullfill the assumption of a gaussian distribution. For alternatives, see documentation of ggpubr::stat_pwc.}

\item{p.adjust.method}{Method for adjusting p-values (see ggpubr::stat_pwc for available methods)}

\item{palette}{Choose a palette of your liking. For available palettes, see ggsci package. Default: "lancet"}

\item{legend.position}{Where to put the legend. Possible options: "top", "right", "bottom", "left"}
}
\value{
Plotting function to show the cell type composition from HiTME object across different samples.
}
\description{
Render box plots summarizing celltype proportions and distribution in samples and groups
}
