% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.aggregated.profile}
\alias{get.aggregated.profile}
\title{Compute aggregated gene expression}
\usage{
get.aggregated.profile(
  object,
  group.by.aggregated = NULL,
  assay = "RNA",
  useNA = FALSE,
  ...
)
}
\arguments{
\item{object}{A seurat object or a list of seurat objects}

\item{group.by.aggregated}{The Seurat object metadata column(s) containing celltype annotations}

\item{assay}{Assay to retrieve information. By default "RNA".}

\item{useNA}{logical whether to return aggregated profile for NA (undefined) cell types, default is FALSE.}

\item{...}{Extra parameters for internal Seurat functions: AverageExpression, AggregateExpression, FindVariableFeatures}
}
\value{
Average and aggregated expression as a list of matrices for all genes and indicated gene lists filtering.
}
\description{
Function to compute aggregated expression (pseudobulk, i.e. sum counts per ident), and average expression by indicated cell type or grouping variable.
}
