% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.GOList}
\alias{get.GOList}
\title{Retrieve list of genes from GO accessions}
\usage{
get.GOList(
  GO_accession = NULL,
  species = "Human",
  host = "https://dec2021.archive.ensembl.org/"
)
}
\arguments{
\item{GO_accession}{Vector of GO accession ID (e.g. \code{c("GO:0003700","GO:0005125")}). Vector could be named for clarity in the output.}

\item{species}{species to retrieve the genes, for now suppported human and mice.}

\item{host}{BioMart host to connect. Default is \link{https://dec2021.archive.ensembl.org/}, as it fails less.}
}
\value{
List of genes for each GO accession requested. If named vector is provided, lists output are named as GO:accession.ID_named (e.g. "GO:0004950_cytokine_receptor_activity")
}
\description{
Function to fetch GO accessions gene list from biomaRt.
}
