% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.HiTObject}
\alias{get.HiTObject}
\title{Get a HiT object summarizing cell type classification and aggregated profiles.}
\usage{
get.HiTObject(
  object,
  group.by = list(layer1 = c("scGate_multi"), layer2 = c("functional.cluster")),
  name.additional.signatures = NULL,
  useNA = FALSE,
  ...
)
}
\arguments{
\item{object}{A seurat object}

\item{group.by}{List with one or multiple Seurat object metadata columns with cell type predictions to group by (e.g. layer 1 cell type classification)}

\item{name.additional.signatures}{Names of additional signatures as found in object metadata to take into account.}

\item{...}{Additional parameters for \link{get.aggregated.profile}, \link{get.celltype.composition}, and \link{get.aggregated.signature} functions.}
}
\value{
HiT object summarizing cell type classification and aggregated profiles.
}
\description{
Get a HiT object summarizing cell type classification and aggregated profiles.
}
