% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{infer.Sex}
\alias{infer.Sex}
\title{Infer the sex of single-cell transcriptomic data based on the expression of X and Y chromosome genes}
\usage{
infer.Sex(
  object = NULL,
  infer.level = c("sample", "cell", "both"),
  split.by = NULL,
  return.Seurat = TRUE,
  ncores = parallel::detectCores() - 2,
  bparam = NULL,
  progressbar = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{Either one or a list of Seurat objects or count matrix (matrix or dcGMatrix)}

\item{infer.level}{Whether to infer sex for cells, samples or both.}

\item{split.by}{Split by sample based on a variable of metadata, either a metadata columns for Seurat object (category) or a vector indicating the sample procedence for each cell}

\item{return.Seurat}{Whether to return Seurat object with additional metadata or dataframe summarizing sex imputation, the latter only is applied if \code{infer.level = "sample"} is indicated.}

\item{ncores}{The number of cores to use, by default all available cores minus 2 are used.}

\item{bparam}{A \link[BiocParallel]{bpparam} object that tells Run.HiTME how to parallelize. If provided, it overrides the \code{ncores} parameter.}

\item{progressbar}{Whether to show a progress bar or not}

\item{verbose}{Verbose output, by default output messsage are returned}
}
\value{
Sample or cell sex inference, returned as extra metadata in Seurat object or as a dataframe with infered sex per sample.
}
\description{
This function takes single-cell (Seurat object or count matrix) transcriptomics data and returns the inferred sex of the cells and individuals samples provided.
Given the sparsity of single-cell RNA-seq data inferring of the sex in many cells is not possible (NA are returned), however, sex inference per sample is highly confident based on pseudobulk data.
\if{html}{\out{<br>}}
Currently only suppoted for human data.
}
