% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{merge.HiTObjects}
\alias{merge.HiTObjects}
\title{Merge HiTObjects}
\usage{
\method{merge}{HiTObjects}(
  hit.object = NULL,
  group.by = NULL,
  metadata.vars = NULL,
  pseudobulk.matrix = "list",
  ncores = parallelly::availableCores() - 2,
  bparam = NULL,
  progressbar = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{hit.object}{List of HiTObjects}

\item{group.by}{If only merging for certain layers of annotation is intended, layers names can be indicated here as vector. Otherwise all layers present in all HiT object will be merged.}

\item{metadata.vars}{Variables to keep as metadata. (Default: NULL, keeping unique metadata columns per sample, dropping single-cell metadata)}

\item{pseudobulk.matrix}{Paramater to determine whther obtain the pseudobulk matrix as a single matrix (\code{"unique"}), or as one matrix for each cell type in the layer (\code{"list"})}

\item{ncores}{The number of cores to use, by default, all available cores - 2.}

\item{bparam}{A \code{BiocParallel::bpparam()} object that tells how to parallelize. If provided, it overrides the \code{ncores} parameter.}

\item{progressbar}{Whether to show a progressbar or not}

\item{verbose}{Whether to show optional messages or not}
}
\value{
Merged HiTObject
}
\description{
Merge HiTObjects
}
