% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{nas.per.sample}
\alias{nas.per.sample}
\title{Render plots summarizing celltype proportions and distribution in samples}
\usage{
nas.per.sample(
  obj.list = NULL,
  annot.col = c("scGate_multi"),
  return.plot = TRUE,
  bottom.mar = 10.2
)
}
\arguments{
\item{obj.list}{List of Seurat objects}

\item{annot.col}{Metadata column(s) containing the cell type annotations}

\item{bottom.mar}{Adjustable bottom margin for long sample names}
}
\value{
Get percentage of not annotated cells per sample and plot it.
}
\description{
Render plots summarizing celltype proportions and distribution in samples
}
