% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.confusion}
\alias{plot.confusion}
\title{Build confusion matrix-like or bar plots between different cell type classification approaches}
\usage{
\method{plot}{confusion}(
  object = NULL,
  var.1 = NULL,
  var.2 = NULL,
  relative = FALSE,
  useNA = "ifany",
  type = "tile",
  xlab = NULL,
  ylab = NULL,
  plot.title = ""
)
}
\arguments{
\item{object}{Seurat object or its metadata with cell type annotation}

\item{var.1}{1st of grouping variables on the x-axis. If using \code{relative = TRUE} proportions will be normalized to this variable.}

\item{var.2}{2nd of grouping variables on the x-axis.}

\item{relative}{Whether to show absolute number of cells, or relative number cells out of the 1st variable indicted. Default is FALSE.}

\item{useNA}{Whether to include not annotated cells or not (labelled as "NA"). Can be "no", "ifany", or "always". See \code{?table} for more information. Default is "ifany".}

\item{type}{Type of plot to render. Either "tile" (default) as a confusion matrices or "barplot".}

\item{xlab}{Label for x-axis}

\item{ylab}{Label for y-axis}

\item{plot.title}{Plot title}
}
\value{
Plots to evaluate the correspondence between different classification methods.
}
\description{
Build confusion matrix-like or bar plots between different cell type classification approaches
}
