% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Run.HiTME}
\alias{Run.HiTME}
\title{Classify cells using scGate and ProjecTILs.}
\usage{
Run.HiTME(
  object = NULL,
  scGate.model = "default",
  scGate.model.branch = c("master", "dev"),
  multi.asNA = TRUE,
  additional.signatures = "default",
  ref.maps = NULL,
  split.by = NULL,
  layer1_link = "CellOntology_ID",
  return.Seurat = TRUE,
  group.by = list(layer1 = c("scGate_multi"), layer2 = c("functional.cluster")),
  useNA = FALSE,
  remerge = TRUE,
  species = "human",
  bparam = NULL,
  ncores = parallelly::availableCores() - 2,
  progressbar = TRUE
)
}
\arguments{
\item{object}{A seurat object or a list of seurat objects}

\item{scGate.model}{The scGate model to use (use get_scGateDB() to get a list of available models), by default fetch the HiTME models: \code{ scGate::get_scGateDB(branch = scGate.model.branch)[[species]][["HiTME"]]}.}

\item{scGate.model.branch}{From which branch Run.HiTME fetch the scGate models, by default models are retrieved from \code{master} branch.}

\item{multi.asNA}{How to label cells that are "Pure" for multiple annotations: "Multi" (FALSE) or NA (TRUE)}

\item{additional.signatures}{UCell additional signatures to compute on each cell, by default \code{SignatuR} programs are included.}

\item{ref.maps}{A named list of the ProjecTILs reference maps to use. They ought to be Seurat objects. It is recommended to add in reference object slost \code{misc} the identifier connecting to layer 1 classification (scGate): \code{ref.map@misc$layer1_link}}

\item{split.by}{A Seurat object metadata column to split by (e.g. sample names).}

\item{layer1_link}{Column of metadata linking layer1 prediction (e.g. scGate ~ CellOntology_ID) in order to perform subsetting for second layer classification.}

\item{return.Seurat}{Whether to return a Hit object or add the data into Seurat object metadata}

\item{group.by}{If return.Seurat = F, variables to be used to summarize HiTME classification data in HiT object. See \link{get.HiTObject} for more information.}

\item{useNA}{Whether to include not annotated cells or not (labelled as "NA" in the group.by.composition) when summarizing into HiT object. Can be "no", "ifany", or "always". See \code{?table} and \link{get.celltype.composition} for more information. Default is "ifany".}

\item{remerge}{When setting split.by, if remerge = TRUE one object will be returned. If remerge = FALSE a list of objects will be returned.}

\item{species}{Define species used for to get correct gene signature list with SignatuR}

\item{bparam}{A \code{BiocParallel::bpparam()} object that tells Run.HiTME how to parallelize. If provided, it overrides the \code{ncores} parameter.}

\item{ncores}{The number of cores to use, by default all available cores minus 2 are used.}

\item{progressbar}{Whether to show a progressbar or not}
}
\value{
Seurat object with additional metadata showing cell type classification, or HiT object summarizing such classification.
}
\description{
Classify cells using scGate and ProjecTILs.
}
\examples{
library(ProjecTILs)

path_data <- file.path("~", "Dropbox", "CSI", "Standardized_SingleCell_Datasets", "ZhangY_2022_34653365", "output", "test")

# Define scGate model
scGate_models_DB <- scGate::get_scGateDB(branch = "master", verbose = TRUE, force_update = TRUE)
models.TME <- scGate_models_DB$human$TME_HiRes

# Load ProjecTILs reference maps
path_ref <- "~/Dropbox/CSI/reference_atlases"
ref.maps <- list(CD8 = load.reference.map(file.path(path_ref, "CD8T_human_ref_v1.rds")),
                 CD4 = load.reference.map(file.path(path_ref, "CD4T_human_ref_v2.rds")),
                 DC = load.reference.map(file.path(path_ref, "DC_human_ref_v1.rds")),
                 MoMac = load.reference.map(file.path(path_ref, "MoMac_human_v1.rds")))
Run.HiTME(object = seurat.object,
          scGate.model = models.TME,
          ref.maps = ref.maps,
          ncores = 6)

}
