% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{save_objs}
\alias{save_objs}
\title{Save object list to disk, in parallel}
\usage{
save_objs(
  obj.list,
  dir,
  ncores = parallelly::availableCores() - 2,
  progressbar = TRUE
)
}
\arguments{
\item{obj.list}{A list of Seurat objects}

\item{dir}{File location}

\item{ncores}{Number of CPU cores to use for parallelization}

\item{progressbar}{Whether to show a progressbar or not}
}
\value{
A list of Seurat objects saved to disk as separate .rds files
}
\description{
Save object list to disk, in parallel
}
\examples{
save_objs(obj.list, "./output/samples")
}
