% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.aggregated.signature}
\alias{get.aggregated.signature}
\title{Compute aggregated additional signatures by cell type}
\usage{
get.aggregated.signature(
  object,
  group.by.aggregated = NULL,
  min.cells.aggregated = 10,
  name.additional.signatures = NULL,
  fun = mean,
  useNA = FALSE
)
}
\arguments{
\item{object}{A seurat object or metadata data frame.}

\item{group.by.aggregated}{The Seurat object metadata column(s) containing celltype annotations (idents).}

\item{min.cells.aggregated}{Minimum number of cells per sample and cell type to aggregate data for pseudobulk or aggregated signatures. Aggregated data for cell types with less than that value will not be returned in aggregated data. Default value is 10.}

\item{name.additional.signatures}{Names of additional signatures to compute the aggregation per cell type.}

\item{fun}{Function to aggregate the signature, e.g. mean or sum.}

\item{useNA}{logical whether to return aggregated signatures for NA (undefined) cell types, default is FALSE.}
}
\value{
Aggregated signature score for each indicated cell type grouping Results is NULL of not additional signatures are indicated or present in metadata.
}
\description{
Function to compute aggregated signatures of predicted cell types.
}
