% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.cluster.samples}
\alias{get.cluster.samples}
\title{Compute aggregation metrics of summarized data}
\usage{
get.cluster.samples(
  object,
  group.by = list(layer1 = c("scGate_multi"), layer2 = c("functional.cluster")),
  min.cells = 10,
  metadata.vars = NULL,
  score = c("silhouette"),
  dist.method = "euclidean",
  ndim = 10,
  nVarGenes = 500,
  gene.filter = NULL,
  GO_accession = NULL,
  ntests = 100,
  black.list = NULL,
  ncores = parallelly::availableCores() - 2,
  bparam = NULL,
  progressbar = TRUE
)
}
\arguments{
\item{object}{List of Hit class object}

\item{group.by}{Grouping variable for}

\item{min.cells}{Minimum number of cells to consider for aggregated expression.}

\item{metadata.vars}{Variables to show as metadata}

\item{score}{Score to compute clustering of samples based on cell type prediction, either Silhoutte or modularity.}

\item{dist.method}{Method to compute distance between celltypes, default euclidean.}

\item{ndim}{Number of dimensions to be use for PCA clustering metrics.}

\item{nVarGenes}{Number of variable genes to assess samples.}

\item{gene.filter}{Additional named list of genes to subset their aggregated expression, if "default" is indicated, all genes,
ribosomal genes, transcription factor (GO:0003700), cytokines (GO:0005125),
cytokine receptors (GO:0004896), chemokines (GO:0008009), and chemokines receptors (GO:0004950) are subsetted and accounted for.}

\item{GO_accession}{Additional GO accessions to subset genes for aggregation, by default the list indicated in \code{gene.filter} are returned.}

\item{black.list}{List of genes to discard from clustering, if "default" object "default_black_list" object is used. Alternative black listed genes can be provided as a vector or list.}

\item{ncores}{The number of cores to use, by default, all available cores - 2.}

\item{bparam}{A \code{BiocParallel::bpparam()} object that tells how to parallelize. If provided, it overrides the \code{ncores} parameter.}

\item{progressbar}{Whether to show a progressbar or not}
}
\value{
List of genes for each GO accession requested. If named vector is provided, lists output are named as GO:accession.ID_named (e.g. "GO:0004950_cytokine_receptor_activity")
}
\description{
Compute aggregation metrics of summarized data
}
