% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.cluster.score}
\alias{get.cluster.score}
\title{Get metrics of cell types pseudobulk clustering}
\usage{
get.cluster.score(
  object = NULL,
  metadata = NULL,
  cluster.by = NULL,
  batching = NULL,
  scores = c("Silhouette_isolated", "Silhouette", "Modularity"),
  dist.method = "euclidean",
  hclust.method = "complete",
  all.concatenated.na.handling = "impute",
  ntests = 100,
  seed = 22,
  pval.combine.method = "weighted_zmethod",
  pca_comps_labs_invisible = c("quali"),
  pca_pb_labs_invisible = c("var", "quali"),
  pca_sig_labs_invisible = c("quali"),
  ndim = 10,
  nVarGenes = 500,
  gene.filter = "HVG",
  black.list = NULL,
  ncores = parallelly::availableCores()/2,
  bparam = NULL,
  progressbar = TRUE
)
}
\arguments{
\item{object}{A Hit class object obtained with \link{get.HiTObject} or pseudobulk raw count matrix (\code{HitObject@aggregated_profile$pseudobulk$layer1})}

\item{metadata}{Metadata data frame corresponding for each sample in the pseudobulk matrix. Row names in this dataframe should be identical as the colnames of the count matrix. If not provided (\code{NULL}) metadata will be extracted from HiT object provided.}

\item{cluster.by}{Vector indicating the variable for clustering, default is celltype (for the annotation) and sample}

\item{batching}{Vector indicating the variable for batching to allow calculating scores per batch, to account for batch effect}

\item{scores}{Scores to compute clustering of samples based on cell type prediction.}

\item{dist.method}{Method to compute distance between celltypes, default is euclidean.}

\item{hclust.method}{Hierarchical clustering method for hclust. Options are: "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC). (See hclust package for details)}

\item{all.concatenated.na.handling}{How to handle missing values in the "all cell types concatenated vector". Options are "drop" (drop rows, conservative method) or "zero_impute" (replaces missing values with zeros, use with caution as it might change your sample clustering based on number of missing values per sample/group)}

\item{ntests}{Number of shuffling events to calculate p-value for scores}

\item{seed}{Set seed for random shuffling events to calculate p-value for scores}

\item{pca_comps_labs_invisible}{Parameter to pass to fviz_pca defining which labels to plot for composition PCA, see documentation of fviz_pca for details.}

\item{pca_pb_labs_invisible}{Parameter to pass to fviz_pca defining which labels to plot for pseudobulk PCA, see documentation of fviz_pca for details.}

\item{pca_sig_labs_invisible}{Parameter to pass to fviz_pca defining which labels to plot for signatures PCA, see documentation of fviz_pca for details.}

\item{ndim}{Number of dimensions to be use for PCA clustering metrics. Default is 10.}

\item{nVarGenes}{Number of variable genes to assess samples. Default is 500.}

\item{gene.filter}{Named list of genes to subset their aggregated expression. Default is \code{"HVG"} or \code{NULL} would indicate "Highly variable genes", number of genes can be set at \code{nVarGenes}. If "default_filter" is indicated transcription factor (GO:0003700), cytokines (GO:0005125), cytokine receptors (GO:0004896), chemokines (GO:0008009), and chemokines receptors (GO:0004950) are subsetted out of the total genes and accounted for. For additional GO accession gene list you may use the function \link{getGOList} to fetch them from biomaRt.}

\item{black.list}{List of genes to discard from clustering, if "default" object "default_black_list" object is used. Alternative black listed genes can be provided as a vector or list.}

\item{ncores}{The number of cores to use, by default, all available cores - 2.}

\item{bparam}{A \code{BiocParallel::bpparam()} object that tells how to parallelize. If provided, it overrides the \code{ncores} parameter.}

\item{progressbar}{Whether to show a progressbar or not}

\item{pval.combine}{Method for combining p-values if calculated using batching. Default is "zmethod" weighted (Stouffer's) Z-method, weighting by sample size per batch. Alternatively, Fisher's method can be used with "fisher".}
}
\value{
Metrics of cell types pseudobulk clustering
}
\description{
Get metrics of cell types pseudobulk clustering
}
