% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{get.HiTObject}
\alias{get.HiTObject}
\title{Get a HiT object summarizing cell type classification and aggregated profiles.}
\usage{
get.HiTObject(
  object,
  group.by = list(layer1 = c("scGate_multi"), layer2 = c("functional.cluster")),
  name.additional.signatures = NULL,
  useNA = FALSE,
  clr_zero_impute_perc = 1,
  layers_links = c(scGate_multi = "functional.cluster"),
  gene.filter = NULL,
  assay = "RNA",
  layer1_link = "CellOntology_ID"
)
}
\arguments{
\item{object}{A Seurat object}

\item{group.by}{List with one or multiple Seurat object metadata columns with cell type predictions to group by (e.g. layer 1 cell type classification)}

\item{name.additional.signatures}{Names of additional signatures as found in object metadata to take into account.}

\item{useNA}{logical whether to return aggregated profile for NA (undefined) cell types, default is FALSE.}

\item{clr_zero_impute_perc}{Parameter for internal \link{get.celltype.composition}.}

\item{layers_links}{Parameter for internal \link{get.celltype.composition}}

\item{gene.filter}{List of genes to subset for aggregated expression. Parameter for internal \link{get.aggregated.profile}.}

\item{assay}{Parameter for internal \link{get.aggregated.profile}.}

\item{layer1_link}{Column of metadata linking layer1 prediction (e.g. scGate ~ CellOntology_ID) in order to perform subsetting for second layer classification.}
}
\value{
HiT object summarizing cell type classification and aggregated profiles.
}
\description{
Get a HiT object summarizing cell type classification and aggregated profiles.
}
