% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{infer.Sex}
\alias{infer.Sex}
\title{Infer the sex of single-cell transcriptomic data based on the expression of X and Y chromosome genes}
\usage{
infer.Sex(
  object = NULL,
  split.by = NULL,
  return.Seurat = TRUE,
  ncores = parallel::detectCores() - 2,
  bparam = NULL,
  progressbar = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{Either one or a list of Seurat objects or count matrix (matrix or dcGMatrix)}

\item{split.by}{Split by sample based on a variable of metadata, either a metadata columns for Seurat object (category) or a vector indicating the sample procedence for each cell}

\item{return.Seurat}{return Seurat object or dataframe summarizing sex imputation. If Seurat object is provided by default Seurat object with additional metadata with infered sex on cell- and sample-wise is returned. If not Seurat object are provided, but matrices, by default dataframe with results are returned}

\item{ncores}{The number of cores to use, by default all available cores minus 2 are used.}

\item{bparam}{A \link[BiocParallel]{bpparam} object that tells Run.HiTME how to parallelize. If provided, it overrides the \code{ncores} parameter.}

\item{progressbar}{Whether to show a progress bar or not}

\item{verbose}{Verbose output, by default output messsage are returned}
}
\value{
List of genes for each GO accession requested. If named vector is provided, lists output are named as GO:accession.ID_named (e.g. "GO:0004950_cytokine_receptor_activity")
}
\description{
This function takes single-cell (Seurat object or count matrix) transcriptomics data and returns the inferred sex of the cells and individuals samples provided.
Given the sparsity of single-cell RNA-seq data inferring of the sex in many cells is not possible (NA are returned), however, sex inference per sample is highly confident based on pseudobulk data.
\if{html}{\out{<br>}}
Currently only suppoted for human data.
}
