% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.celltype.freq}
\alias{plot.celltype.freq}
\title{Render plots summarizing celltype proportions and distribution in samples}
\usage{
\method{plot}{celltype.freq}(
  object = NULL,
  group.by = list(layer1 = c("scGate_multi"), layer2 = c("functional.cluster")),
  split.by = NULL,
  by.x = "celltype"
)
}
\arguments{
\item{object}{List of Hit class object}

\item{group.by}{Grouping variable for}

\item{split.by}{If desired, indicate how to split plots by metadata variable.}

\item{by.x}{Determine which variable show on x-axis, if celltype it shows boxplots of the number of each cell type for each sample. If sample is indicated a barplot is shown with the relative proportions of each celltype within each sample.}
}
\value{
List of genes for each GO accession requested. If named vector is provided, lists output are named as GO:accession.ID_named (e.g. "GO:0004950_cytokine_receptor_activity")
}
\description{
Render plots summarizing celltype proportions and distribution in samples
}
