% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{read_objs}
\alias{read_objs}
\title{Read all .rds files in a directory and return list of Seurat objects, in parallel}
\usage{
read_objs(
  dir = NULL,
  file.list = NULL,
  ncores = parallelly::availableCores() - 2,
  progressbar = TRUE
)
}
\arguments{
\item{dir}{File location}

\item{file.list}{A list of files (with full pathname)}

\item{ncores}{Number of CPU cores to use for parallelization}

\item{progressbar}{Whether to show a progressbar or not}
}
\value{
A list of Seurat objects read into R
}
\description{
Read all .rds files in a directory and return list of Seurat objects, in parallel
}
\examples{
obj.list <- read_objs("./output/samples")
}
