% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Run.HiTME}
\alias{Run.HiTME}
\title{Annotate cell types at multiple levels of granularity in a scRNA-seq dataset.}
\usage{
Run.HiTME(
  object = NULL,
  scGate.model = "default",
  scGate.model.branch = c("master", "dev"),
  multi.asNA = TRUE,
  additional.signatures = NULL,
  ref.maps = NULL,
  layer3 = "default",
  layer3.threshold = 0.2,
  split.by = NULL,
  layer1_link = "CellOntology_ID",
  remerge = TRUE,
  species = "human",
  bparam = NULL,
  ncores = parallel::detectCores() - 2,
  progressbar = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{A seurat object or a list of seurat objects}

\item{scGate.model}{The \link[scGate]{scGate} model to use. Use \link[scGate]{get_scGateDB} to get a list of available models. By default it fetchs the HiTME models: \code{ scGate::get_scGateDB(branch = scGate.model.branch)[[species]][["HiTME"]]}}

\item{scGate.model.branch}{From which branch Run.HiTME fetch the scGate models, by default models are retrieved from \code{master} branch.}

\item{multi.asNA}{How to label cells that are "Pure" for multiple annotations: "Multi" (FALSE) or NA (TRUE)}

\item{additional.signatures}{Additional signature(s) to compute on each cell using \link[UCell]{UCell}. Multiple signatures should be provided as a list.}

\item{ref.maps}{A named list of the \link\link{ProjecTILs}{ProjecTILs} reference maps to use. Download default reference maps using \link\href{get.reference.maps}{ProjecTILs}. For custom reference maps, they must be Seurat objects turn into reference with \link\href{make.reference}{ProjecTILs}. It is recommended to add in reference object slost \code{misc} the identifier connecting to layer 1 classification (scGate): \code{ref.map@misc$layer1_link}}

\item{layer3}{Gene signature(s) to define gene programs or cell states on top of layer2 (\link[ProjecTILs]{ProjecTILs} classification). By default \link[SignatuR]{SignatuR} programs for cell cycling, IFN response and HeatShock response are run.}

\item{layer3.threshold}{Minimum value threshold to apply on \link[UCell]{UCell} scores of layer3 signatures to classify cell types.}

\item{split.by}{A Seurat object metadata column to split by (e.g. sample names).}

\item{layer1_link}{Column of metadata linking layer1 prediction (e.g. scGate ~ CellOntology_ID) in order to perform subsetting for second layer classification.}

\item{remerge}{When setting split.by or providing a list of objects, if \code{remerge = TRUE} one object will be returned (default). If \code{remerge = FALSE} a list of objects will be returned.}

\item{species}{Define species to get the default \link[scGate]{scGate} models and \link[SignatuR]{SignatuR} signatures. Currently only human and mouse are supported, if other species are studied, set to \code{NULL}}

\item{bparam}{A \link[BiocParallel]{bpparam} object that tells Run.HiTME how to parallelize. If provided, it overrides the \code{ncores} parameter.}

\item{ncores}{The number of cores to use, by default all available cores minus 2 are used.}

\item{progressbar}{Whether to show a progress bar or not}

\item{verbose}{Verbose output, by default output messsage are returned}
}
\value{
Seurat object with additional metadata showing cell type classification.
}
\description{
The function takes as input Seurat objects (or list of them).
These should be split by sample to avoid batch effects, or split internally in by indicating the parameter \code{split.by}.
\if{html}{\out{<br>}}
This function firstly runs \link[scGate]{scGate} (easily customizable) marker-based classification, resulting in a coarse-grained cell type classification (CD4T, B cell, Dendritic cell...).
Next, it runs for each broad cell type \link[ProjecTILs]{ProjecTILs} for a finer cell type classification (CD4+ TFH, Tex CD8+, cDC1...) based on cell mapping onto expert-curated single-cell reference maps.
}
\examples{
library(ProjecTILs)
library(HiTME)

# get reference maps for human
ref.maps <- get.reference.maps(collection  = "human")

# Run HiTME with default settings, specifying the human ref.maps
# Internally default scGate models and gene programs signatures will be fetched

 query <- Run.HiTME(object = query,
                    ref.maps = ref.maps[["human"]]
                    )

}
