% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Run.HiTME}
\alias{Run.HiTME}
\title{Run HitME: classify cells}
\usage{
Run.HiTME(
  object = NULL,
  scGate.model = "default",
  scGate.model.branch = c("master", "dev"),
  additional.signatures = "default",
  ref.maps = NULL,
  split.by = NULL,
  layer1_link = "CellOntology_ID",
  return.Seurat = TRUE,
  group.by = list(layer1 = c("scGate_multi"), layer2 = c("functional.cluster")),
  useNA = FALSE,
  remerge = TRUE,
  species = "human",
  bparam = NULL,
  ncores = parallelly::availableCores() - 2,
  progressbar = TRUE
)
}
\arguments{
\item{object}{A seurat object or a list of seurat objects}

\item{scGate.model}{The scGate model to use (use get_scGateDB() to get a list of available models), by default fetch the HiTME models: \code{ scGate::get_scGateDB(branch = scGate.model.branch,)[[species]][["HiTME"]]}.}

\item{scGate.model.branch}{From which branch Run.HiTME fetch the scGate models, by default models are retrieved from \code{master} branch.}

\item{additional.signatures}{UCell additional signatures to compute on each cell, by default \code{SignatuR} Programs are included.}

\item{ref.maps}{A named list of the ProjecTILs reference maps to use.}

\item{split.by}{A Seurat object metadata column to split by (e.g. sample names).}

\item{layer1_link}{Column of metadata linking layer1 prediction in order to perform subsetting for second layer classification.}

\item{return.Seurat}{Whether to return a Hit object or add the data into Seurat object metadata}

\item{group.by}{If return.Seurat = F, variables to be used to summarize HiTME classification data in HiT object.}

\item{useNA}{Whether to include not annotated cells or not (labelled as "NA" in the group.by.composition) when summarizing into HiT object.}

\item{remerge}{When setting split.by, if remerge = TRUE one object will be returned. If remerge = FALSE a list of objects will be returned.}

\item{bparam}{A \code{BiocParallel::bpparam()} object that tells Run.HiTME how to parallelize. If provided, it overrides the \code{ncores} parameter.}

\item{ncores}{The number of cores to use}

\item{progressbar}{Whether to show a progressbar or not}
}
\value{
No return in R. The input files will be overwritten.
}
\description{
Run HitME: classify cells
}
\examples{
library(ProjecTILs)

path_data <- file.path("~", "Dropbox", "CSI", "Standardized_SingleCell_Datasets", "ZhangY_2022_34653365", "output", "test")

# Define scGate model
scGate_models_DB <- scGate::get_scGateDB(branch = "master", verbose = T, force_update = TRUE)
models.TME <- scGate_models_DB$human$TME_HiRes

# Load ProjecTILs reference maps
path_ref <- "~/Dropbox/CSI/reference_atlases"
ref.maps <- list(CD8 = load.reference.map(file.path(path_ref, "CD8T_human_ref_v1.rds")),
                 CD4 = load.reference.map(file.path(path_ref, "CD4T_human_ref_v2.rds")),
                 DC = load.reference.map(file.path(path_ref, "DC_human_ref_v1.rds")),
                 MoMac = load.reference.map(file.path(path_ref, "MoMac_human_v1.rds")))
Run.HiTME(object = seurat.object,
          scGate.model = models.TME,
          ref.maps = ref.maps,
          ncores = 6)

}
